requirements.txt
django
django-reversion
django-tagging
django-admin-sortable
djangorestframework